#1234 updated documentation, javadoc, license header and CLI command name, including refactoring of the main entry class names to reflect the command line name
