+++ deegree-tools/deegree-tools-gml/src/main/java/org/deegree/tools/featurestoresql/config/SqlFeatureStoreConfigCreatorApp.java
@@ -19,7 +19,7 @@
  * <http://www.gnu.org/licenses/lgpl-2.1.html>.
  * #L%
  */
-package org.deegree.tools.featurestoresql.loader;
+package org.deegree.tools.featurestoresql.config;
 
 import org.springframework.boot.Banner;
 import org.springframework.boot.SpringApplication;
@@ -27,28 +27,29 @@ import org.springframework.boot.SpringApplication;
 import org.deegree.tools.featurestoresql.CommonConfiguration;
 
 /**
- * Entry point of the command line interface of GmlLoader.
+ * Entry point of the command line interface of SqlFeatureStoreConfigCreator.
  *
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
  */
-public class GmlLoaderApp {
+public class SqlFeatureStoreConfigCreatorApp {
 
     public static void run( String[] args ) {
         if ( args.length == 1
              || ( args.length > 1 && ( "--help".equals( args[1] ) || "-help".equals( args[1] ) || "-h".equals( args[1] ) ) ) ) {
-            GmlLoaderHelpUsage.printUsage();
-        } else if ( args.length < 4 ) {
-            printUnexpectedNumberOfParameters( args );
-            GmlLoaderHelpUsage.printUsage();
+            SqlFeatureStoreConfigCreatorUsagePrinter.printUsage();
+        } else if ( args.length == 1 ) {
+            printUnexpectedNumberOfParameters();
+            SqlFeatureStoreConfigCreatorUsagePrinter.printUsage();
         } else {
-            SpringApplication app = new SpringApplication( CommonConfiguration.class, GmlLoaderConfiguration.class );
+            SpringApplication app = new SpringApplication( CommonConfiguration.class,
+                                                           SqlFeatureStoreConfigCreatorConfiguration.class );
             app.setBannerMode( Banner.Mode.OFF );
             app.run( args );
         }
     }
 
-    private static void printUnexpectedNumberOfParameters( String[] args ) {
-        System.out.println( "Number of arguments is invalid, must be at least three but was " + ( args.length -1 ) );
+    private static void printUnexpectedNumberOfParameters() {
+        System.out.println( "Number of arguments is invalid, must be more one." );
         System.out.println();
     }
