+++ deegree-tools/deegree-tools-gml/src/main/java/org/deegree/tools/featurestoresql/config/SqlFeatureStoreConfigCreatorConfiguration.java
@@ -1,3 +1,24 @@
+/*-
+ * #%L
+ * deegree-cli-utility
+ * %%
+ * Copyright (C) 2016 - 2021 lat/lon GmbH
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 2.1 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Lesser Public License for more details.
+ *
+ * You should have received a copy of the GNU General Lesser Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ * #L%
+ */
 package org.deegree.tools.featurestoresql.config;
 
 import org.deegree.feature.types.AppSchema;
@@ -20,7 +41,7 @@ import org.springframework.context.annotation.Configuration;
  */
 @Configuration
 @EnableBatchProcessing
-public class FeatureStoreLoaderConfiguration {
+public class SqlFeatureStoreConfigCreatorConfiguration {
 
     @Autowired
     private JobBuilderFactory jobBuilderFactory;
