+++ deegree-tools/deegree-tools-gml/src/main/java/org/deegree/tools/featurestoresql/config/SqlFeatureStoreConfigCreatorUsagePrinter.java
@@ -1,12 +1,35 @@
+/*-
+ * #%L
+ * deegree-cli-utility
+ * %%
+ * Copyright (C) 2016 - 2021 lat/lon GmbH
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 2.1 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Lesser Public License for more details.
+ *
+ * You should have received a copy of the GNU General Lesser Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ * #L%
+ */
 package org.deegree.tools.featurestoresql.config;
 
 /**
+ * SqlFeatureStoreConfigCreator CLI usage info.
+ *
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
  */
-public class FeatureStoreConfigUsagePrinter {
+public class SqlFeatureStoreConfigCreatorUsagePrinter {
 
     public static void printUsage() {
-        System.out.println( "Usage: java -jar deegree-gml-tool.jar FeatureStoreConfigLoader -schemaUrl=<url-or-path/to/file> [options]" );
+        System.out.println( "Usage: java -jar deegree-tools-gml.jar SqlFeatureStoreConfigCreator -schemaUrl=<url-or-path/to/file> [options]" );
         System.out.println();
         System.out.println( "arguments:" );
         System.out.println( " -schemaUrl=<url-or-path/to/file>, path to the schema, may be an local reference or http url" );
