+++ deegree-tools/deegree-tools-gml/src/main/resources/logback.xml
@@ -1,6 +1,5 @@
 <configuration>
 
-
     <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
         <encoder>
             <pattern>%d{yyyy-MM-dd HH:mm:ss} [%t] %level %logger{36} - %msg%n</pattern>
